refactor(widget): review fixes
