+++ packages/widget-playground/src/components/Mock/MockElement.tsx
@@ -2,12 +2,14 @@ import { Box } from '@mui/material'
 import { styled } from '@mui/material/styles'
 import type React from 'react'
 
+export const mockElementHeight = 48
+
 export const MockElement: React.FC<React.ComponentProps<typeof Box>> = styled(
   Box
 )(({ theme }) => ({
   backgroundColor: theme.vars.palette.background.paper,
   width: '100%',
-  height: 48,
+  height: mockElementHeight,
   display: 'flex',
   justifyContent: 'center',
   alignItems: 'center',
